More of a proper way to implement it
