--- Comparison-shopping-engine-backend/App_Code/CoreEntities/ItemComparers/ItemDateComparer.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-/// <summary>
-/// This class describes the comparison by date logic for Item classes
-/// </summary>
-public class ItemDateComparer : IComparer<Item>
-{
-    public int Compare(Item x, Item y)
-    {
-        return x.Date.CompareTo(y.Date);
-    }
-}
