--- Comparison-shopping-engine-backend/App_Code/CoreEntities/ItemComparers/ItemPriceComparer.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-/// <summary>
-/// This class describes the comparison by price logic for Item classes
-/// </summary>
-public class ItemPriceComparer : IComparer<Item>
-{
-    public int Compare(Item x, Item y)
-    {
-        return x.Price.CompareTo(y.Price);
-    }
-}
