--- Comparison-shopping-engine-backend/App_Code/IService.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using System.Runtime.Serialization;
-using System.ServiceModel;
-using System.ServiceModel.Web;
-using System.Text;
-
-[ServiceContract]
-public interface IService
-{
-    [OperationContract]
-    Receipt TestMethod();
-
-    [OperationContract]
-    Receipt ProcessImage(Bitmap source);
-
-    [OperationContract]
-    List<Item> ProcessReceipt(Receipt source);
-}
