+++ Comparison-shopping-engine-backend/Comparison-shopping-engine-backend/norm/names.list
@@ -1,0 +1,1 @@
+["Ledai Maxima Favorit", "KELMES pienas, 2,5% rieb.", "Malta kava Paulig Extra"]
