+++ Comparison-shopping-engine-backend/Comparison-shopping-engine-backend/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Tesseract" version="3.0.2.0" targetFramework="net452" />
+</packages>
