--- Comparison-shopping-engine-backend/Web.config
@@ -1,34 +1,0 @@
-﻿<?xml version="1.0"?>
-<configuration>
-  <appSettings>
-    <add key="aspnet:UseTaskFriendlySynchronizationContext" value="true"/>
-  </appSettings>
-  <system.web>
-    <compilation debug="true" targetFramework="4.5.2"/>
-    <httpRuntime targetFramework="4.5.2"/>
-  </system.web>
-  <system.serviceModel>
-    <behaviors>
-      <serviceBehaviors>
-        <behavior>
-          <!-- To avoid disclosing metadata information, set the values below to false before deployment -->
-          <serviceMetadata httpGetEnabled="true" httpsGetEnabled="true"/>
-          <!-- To receive exception details in faults for debugging purposes, set the value below to true.  Set to false before deployment to avoid disclosing exception information -->
-          <serviceDebug includeExceptionDetailInFaults="false"/>
-        </behavior>
-      </serviceBehaviors>
-    </behaviors>
-    <protocolMapping>
-      <add binding="basicHttpsBinding" scheme="https"/>
-    </protocolMapping>
-    <serviceHostingEnvironment aspNetCompatibilityEnabled="true" multipleSiteBindingsEnabled="true"/>
-  </system.serviceModel>
-  <system.webServer>
-    <modules runAllManagedModulesForAllRequests="true"/>
-    <!--
-        To browse web app root directory during debugging, set the value below to true.
-        Set to false before deployment to avoid disclosing web app folder information.
-      -->
-    <directoryBrowse enabled="true"/>
-  </system.webServer>
-</configuration>
