--- Comparison-shopping-engine-backend/db/test1.item
@@ -1,1 +1,0 @@
-﻿{"Saved":true,"Name":"KELMES pienas, 2,5% rieb.","Store":"IKI","Price":10,"Date":"\/Date(1253059200000)\/"}
