--- Comparison-shopping-engine-backend/db/test2.item
@@ -1,1 +1,0 @@
-﻿{"Saved":true,"Name":"Malta kava Paulig Extra“","Store":"MAXIMA","Price":899,"Date":"\/Date(1253059200000)\/"}
