--- Comparison-shopping-engine-backend/db/test3.item
@@ -1,1 +1,0 @@
-﻿{"Saved":true,"Name":"Ledai Maxima Favorit","Store":"MAXIMA","Price":129,"Date":"\/Date(1253059200000)\/"}
