--- Comparison-shopping-engine-backend/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Tesseract" version="3.0.2.0" targetFramework="net452" />
-</packages>
