+++ Comparison-shopping-engine/Comparison-shopping-engine.sln
@@ -7,27 +7,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Comparison-shopping-engine"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Comparison-shopping-engine.Tests", "..\Comparison-shopping-engine.Tests\Comparison-shopping-engine.Tests.csproj", "{1DC1F5C2-780E-4EA1-BAC6-7B0DFEF9AA00}"
 EndProject
-Project("{E24C65DC-7377-472B-9ABA-BC803B73C61A}") = "Comparison-shopping-engine-backend", "..\Comparison-shopping-engine-backend\", "{C682AC1B-FAAE-4F8B-B7CD-4DACCB5DFA33}"
-	ProjectSection(WebsiteProperties) = preProject
-		TargetFrameworkMoniker = ".NETFramework,Version%3Dv4.5.2"
-		Debug.AspNetCompiler.VirtualPath = "/localhost_6992"
-		Debug.AspNetCompiler.PhysicalPath = "..\Comparison-shopping-engine-backend\"
-		Debug.AspNetCompiler.TargetPath = "PrecompiledWeb\localhost_6992\"
-		Debug.AspNetCompiler.Updateable = "true"
-		Debug.AspNetCompiler.ForceOverwrite = "true"
-		Debug.AspNetCompiler.FixedNames = "false"
-		Debug.AspNetCompiler.Debug = "True"
-		Release.AspNetCompiler.VirtualPath = "/localhost_6992"
-		Release.AspNetCompiler.PhysicalPath = "..\Comparison-shopping-engine-backend\"
-		Release.AspNetCompiler.TargetPath = "PrecompiledWeb\localhost_6992\"
-		Release.AspNetCompiler.Updateable = "true"
-		Release.AspNetCompiler.ForceOverwrite = "true"
-		Release.AspNetCompiler.FixedNames = "false"
-		Release.AspNetCompiler.Debug = "False"
-		VWDPort = "6992"
-		SlnRelativePath = "..\Comparison-shopping-engine-backend\"
-		DefaultWebSiteLanguage = "Visual C#"
-	EndProjectSection
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Comparison-shopping-engine-backend", "..\Comparison-shopping-engine-backend\Comparison-shopping-engine-backend\Comparison-shopping-engine-backend.csproj", "{657981FE-B065-4537-8B59-7A86B1AD4E5F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,10 +23,10 @@ Global
 		{1DC1F5C2-780E-4EA1-BAC6-7B0DFEF9AA00}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1DC1F5C2-780E-4EA1-BAC6-7B0DFEF9AA00}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1DC1F5C2-780E-4EA1-BAC6-7B0DFEF9AA00}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C682AC1B-FAAE-4F8B-B7CD-4DACCB5DFA33}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C682AC1B-FAAE-4F8B-B7CD-4DACCB5DFA33}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C682AC1B-FAAE-4F8B-B7CD-4DACCB5DFA33}.Release|Any CPU.ActiveCfg = Debug|Any CPU
-		{C682AC1B-FAAE-4F8B-B7CD-4DACCB5DFA33}.Release|Any CPU.Build.0 = Debug|Any CPU
+		{657981FE-B065-4537-8B59-7A86B1AD4E5F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{657981FE-B065-4537-8B59-7A86B1AD4E5F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{657981FE-B065-4537-8B59-7A86B1AD4E5F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{657981FE-B065-4537-8B59-7A86B1AD4E5F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
